+++ TinkoffASDKUI/TinkoffASDKUI/Modules/MainForm/Models/MainFormRowType.swift
@@ -1,0 +1,12 @@
+//
+//  MainFormRowType.swift
+//  TinkoffASDKUI
+//
+//  Created by r.akhmadeev on 24.01.2023.
+//
+
+import Foundation
+
+enum MainFormRowType {
+    case savedCard(ISavedCardViewOutput)
+}