+++ TinkoffASDKUI/TinkoffASDKUI/Modules/MainForm/Presenter/MainFormPresenter.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import TinkoffASDKCore
 
 final class MainFormPresenter {
     // MARK: Dependencies
@@ -13,6 +14,14 @@ final class MainFormPresenter {
     weak var view: IMainFormViewController?
     private let router: IMainFormRouter
     private let stub: MainFormStub
+
+    // MARK: Child Presenters
+
+    private lazy var savedCardPresenter = SavedCardPresenter(output: self)
+
+    // MARK: State
+
+    private var rows: [MainFormRowType] { [.savedCard(savedCardPresenter)] }
 
     // MARK: Init
 
@@ -26,6 +35,19 @@ final class MainFormPresenter {
 
 extension MainFormPresenter: IMainFormPresenter {
     func viewDidLoad() {
+        let paymentCard = PaymentCard(
+            pan: "2201382000000039",
+            cardId: "123456",
+            status: .active,
+            parentPaymentId: nil,
+            expDate: "1030"
+        )
+
+        savedCardPresenter.presentationState = .selected(
+            card: paymentCard,
+            hasAnotherCards: true
+        )
+
         let orderDetails = MainFormOrderDetailsViewModel(
             amountDescription: "К оплате",
             amount: "10 500 ₽",
@@ -42,11 +64,37 @@ extension MainFormPresenter: IMainFormPresenter {
         )
 
         view?.updateHeader(with: header)
+        view?.set(payButtonEnabled: savedCardPresenter.isValid)
     }
 
     func viewWasClosed() {}
 
     func viewDidTapPayButton() {
         router.openCardPaymentForm()
+    }
+
+    func numberOfRows() -> Int {
+        rows.count
+    }
+
+    func row(at indexPath: IndexPath) -> MainFormRowType {
+        rows[indexPath.row]
+    }
+}
+
+// MARK: - ISavedCardPresenterOutput
+
+extension MainFormPresenter: ISavedCardPresenterOutput {
+    func savedCardPresenter(
+        _ presenter: SavedCardPresenter,
+        didRequestReplacementFor paymentCard: PaymentCard
+    ) {}
+
+    func savedCardPresenter(
+        _ presenter: SavedCardPresenter,
+        didUpdateCVC cvc: String,
+        isValid: Bool
+    ) {
+        view?.set(payButtonEnabled: isValid)
     }
 }